+++ spec/cattri/instance_attributes_spec.rb
@@ -26,6 +26,16 @@ RSpec.describe Cattri::InstanceAttributes do
       expect(instance).to respond_to(:items)
     end
 
+    it "defines multiple write-only attributes" do
+      test_class.instance_attribute :x, :y
+      instance = test_class.new
+
+      expect(instance).to respond_to(:x=)
+      expect(instance).to respond_to(:y=)
+      expect(instance).to respond_to(:x)
+      expect(instance).to respond_to(:y)
+    end
+
     it "defines a reader" do
       expect(instance.items).to eq([])
     end
@@ -56,6 +66,12 @@ RSpec.describe Cattri::InstanceAttributes do
         test_class.iattr :bar, default: 10
       end.to raise_error(Cattri::AttributeDefinitionError, /Failed to define method :bar on/)
     end
+
+    it "raises AmbiguousBlockError when using a block with multiple attributes" do
+      expect do
+        test_class.iattr(:foo, :bar) { |v| v }
+      end.to raise_error(Cattri::AmbiguousBlockError, "Cannot define multiple attributes with a block")
+    end
   end
 
   describe ".instance_attribute_reader / .iattr_reader" do
@@ -70,6 +86,15 @@ RSpec.describe Cattri::InstanceAttributes do
     it "does not allow setting a readonly attribute" do
       expect { instance.readonly = "fail" }.to raise_error(NoMethodError)
     end
+
+    it "defines multiple read-only attributes" do
+      test_class.iattr_reader :alpha, :beta, default: "readable"
+
+      expect(instance).to respond_to(:alpha)
+      expect(instance).to respond_to(:beta)
+      expect(instance).not_to respond_to(:alpha=)
+      expect(instance).not_to respond_to(:beta=)
+    end
   end
 
   describe ".instance_attribute_writer / .iattr_writer" do
@@ -83,6 +108,15 @@ RSpec.describe Cattri::InstanceAttributes do
 
     it "does not allow getting a write-only attribute" do
       expect { instance.secret }.to raise_error(NoMethodError)
+    end
+
+    it "defines multiple write-only attributes" do
+      test_class.iattr_writer :x, :y
+
+      expect(instance).to respond_to(:x=)
+      expect(instance).to respond_to(:y=)
+      expect(instance).not_to respond_to(:x)
+      expect(instance).not_to respond_to(:y)
     end
   end